# docs/api/openapi/orders.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Store Order Service",
    "description": "Store Order Service API documentation",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/admin/orders/{id}": {
      "get": {
        "tags": [
          "AdminOrders"
        ],
        "summary": "One order; 404 when the id is unknown.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error, as an RFC 9457 problem (application/problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No valid access token"
          },
          "403": {
            "description": "The signed-in user may not do this"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/v1/orders/pricing-rules": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "The rules an order is priced by (delivery fee and its free-delivery threshold, first-order\ndiscount), for the cart page to preview the amounts the way the checkout will compute them,\nand the days an order placed now should arrive between.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingRulesResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error, as an RFC 9457 problem (application/problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/discount-codes/{code}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "What a discount code takes off subtotal, for the cart to preview; a\ncode that cannot be used (unknown, expired, used up, below its minimum) is 422 saying why.\nThe order checks the code again when it is placed.",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subtotal",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscountCodeCheckResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error, as an RFC 9457 problem (application/problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/from-cart": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Places an order from the cart. Send an Idempotency-Key header (a UUID per checkout\nattempt): a retry with the same key returns the order created the first time.",
        "parameters": [
          {
            "name": "idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderFromCartRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error, as an RFC 9457 problem (application/problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No valid access token"
          },
          "403": {
            "description": "The signed-in user may not do this"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "One order: the customer's own, or any order for an administrator (masked for the demo one).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error, as an RFC 9457 problem (application/problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No valid access token"
          },
          "403": {
            "description": "The signed-in user may not do this"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/orders/{id}/payment": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Opens the payment of the customer's order once its stock is reserved; calling it again gives\nthe same payment. 409 while the stock is still being reserved (try again), once the order is\npaid or cancelled, or after its payment deadline.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPaymentResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error, as an RFC 9457 problem (application/problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No valid access token"
          },
          "403": {
            "description": "The signed-in user may not do this"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/orders/my-orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "The customer's orders, newest first.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponsePagedResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error, as an RFC 9457 problem (application/problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No valid access token"
          },
          "403": {
            "description": "The signed-in user may not do this"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/orders/has-orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Whether the customer has ordered before - the first order is discounted.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HasOrdersResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error, as an RFC 9457 problem (application/problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No valid access token"
          },
          "403": {
            "description": "The signed-in user may not do this"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/webhooks/payments": {
      "post": {
        "tags": [
          "PaymentWebhooks"
        ],
        "summary": "A payment event, signed in the Store-Signature header (HMAC-SHA256 over \"t.body\"). A wrong,\nmissing or stale (older than PaymentWebhooks:ToleranceSeconds) signature is 401; an event\nalready processed is acknowledged again and changes nothing.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentWebhookEvent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookReceipt"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "default": {
            "description": "Error, as an RFC 9457 problem (application/problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateOrderFromCartRequest": {
        "required": [
          "customerName",
          "saveAddress"
        ],
        "type": "object",
        "properties": {
          "deliveryAddress": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "discountCode": {
            "type": "string",
            "description": "A discount code the customer typed; the order is refused when it cannot be used.",
            "nullable": true
          },
          "saveAddress": {
            "type": "boolean",
            "description": "Also store the delivery address in the user profile."
          }
        },
        "additionalProperties": false,
        "description": "Body of POST /api/v1/orders/from-cart. Rules: `CreateOrderFromCartRequestValidator`."
      },
      "DiscountCodeCheckResponse": {
        "required": [
          "code",
          "discountAmount",
          "kind",
          "value"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "value": {
            "type": "number",
            "format": "double"
          },
          "discountAmount": {
            "type": "number",
            "description": "The amount the code takes off the subtotal it was checked with.",
            "format": "double"
          }
        },
        "additionalProperties": false,
        "description": "What a code takes off a subtotal, for the cart to preview before the checkout."
      },
...
      "HasOrdersResponse": {
        "required": [
          "hasOrders",
          "ordersCount"
        ],
        "type": "object",
        "properties": {
          "hasOrders": {
            "type": "boolean"
          },
          "ordersCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "OrderItemResponse": {
        "required": [
          "color",
          "company",
          "id",
          "lineTotal",
          "price",
          "productId",
          "productImage",
          "productTitle",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "productTitle": {
            "type": "string"
          },
          "productImage": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "color": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "lineTotal": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "OrderPaymentResponse": {
        "required": [
          "amount",
          "currency",
          "orderId",
          "paymentDueAt",
          "paymentId",
          "status"
        ],
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "paymentId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "description": "ISO code, lowercase."
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "paymentDueAt": {
            "type": "string",
            "description": "When the order is cancelled if it is still unpaid.",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "The payment of an order, for the payment page: the browser confirms it with a card straight at\nthe payment service (POST /api/v1/payments/{paymentId}/confirm) before Store.OrderService.DTOs.Responses.OrderPaymentResponse.PaymentDueAt."
      },
      "OrderResponse": {
        "required": [
          "createdAt",
          "customerName",
          "deliveryFee",
          "discountAmount",
          "id",
          "nextStatuses",
          "orderItems",
          "status",
          "statusHistory",
          "subtotal",
          "total",
          "totalItems",
          "userEmail",
          "userId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string"
          },
          "userEmail": {
            "type": "string"
          },
          "deliveryAddress": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string"
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemResponse"
            }
          },
          "totalItems": {
            "type": "integer",
            "format": "int32"
          },
          "subtotal": {
            "type": "number",
            "description": "Sum of the lines before discount and delivery.",
            "format": "double"
          },
          "discountAmount": {
            "type": "number",
            "format": "double"
          },
          "discountReason": {
            "type": "string",
            "nullable": true
          },
          "discountCode": {
            "type": "string",
            "description": "The discount code the order was discounted by, if any.",
            "nullable": true
          },
          "deliveryFee": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "description": "What the customer pays.",
            "format": "double"
          },
          "status": {
            "type": "string"
          },
          "statusHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderStatusChangeResponse"
            },
            "description": "Every status the order has been in, oldest first, with the time it changed."
          },
          "nextStatuses": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "What the administrator can still do with the order: Shipped (once paid), Cancelled (until shipped)."
          },
          "deliveryFrom": {
            "type": "string",
            "description": "First and last day of the delivery window promised at checkout; null for older orders.",
            "format": "date",
            "nullable": true
          },
          "deliveryTo": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "paymentDueAt": {
            "type": "string",
            "description": "When the customer must have paid by; set once the stock is reserved.",
            "format": "date-time",
            "nullable": true
          },
          "cardBrand": {
            "type": "string",
            "description": "Brand of the card that paid (visa, mastercard).",
            "nullable": true
          },
          "cardLast4": {
            "type": "string",
            "description": "Last four digits of that card.",
            "nullable": true
          },
          "cancellationReason": {
            "type": "string",
            "description": "Why the order was cancelled: out-of-stock, payment-timed-out or cancelled-by-administrator.",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrderResponsePagedResponse": {
        "required": [
          "hasNextPage",
          "hasPreviousPage",
          "items",
          "page",
          "pageSize",
          "totalCount",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderResponse"
            }
          },
          "totalCount": {
            "type": "integer",
            "description": "Rows in the whole listing, not on this page.",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "hasNextPage": {
            "type": "boolean",
            "readOnly": true
          },
          "hasPreviousPage": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "description": "One page of a listing, the same shape for every resource."
      },
...
      "OrderStatusChangeResponse": {
        "required": [
          "changedAt",
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "changedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "PaymentStatus": {
        "enum": [
          "requiresPaymentMethod",
          "requiresAction",
          "succeeded",
          "cancelled",
          "refunded"
        ],
        "type": "string",
        "description": "Where a payment stands; stored by name, written in camelCase by the APIs."
      },
...
      "PaymentWebhookEvent": {
        "required": [
          "created",
          "data",
          "id",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique per event; the same when a delivery is retried",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "description": "One of Store.Contracts.Payments.Webhooks.PaymentWebhookTypes"
          },
          "created": {
            "type": "string",
            "description": "When the event happened (UTC)",
            "format": "date-time"
          },
          "data": {
            "$ref": "#/components/schemas/PaymentWebhookData"
          }
        },
        "additionalProperties": false,
        "description": "What the payment service posts to the order service's webhook when a payment changes, the way\na card provider reports to a shop: JSON in camelCase, signed in the `Store-Signature`\nheader (`WebhookSignature`). An event may arrive more than once and out of order; its\nId tells a repeat from a new one."
      },
      "PricingRulesResponse": {
        "required": [
          "deliveryFee",
          "deliveryFrom",
          "deliveryTo",
          "firstOrderDiscountPercent",
          "freeDeliveryThreshold"
        ],
        "type": "object",
        "properties": {
          "freeDeliveryThreshold": {
            "type": "number",
            "description": "Orders with a subtotal (before the discount) at or above this amount ship for free.",
            "format": "double"
          },
          "deliveryFee": {
            "type": "number",
            "description": "Charged below the threshold.",
            "format": "double"
          },
          "firstOrderDiscountPercent": {
            "type": "number",
            "description": "Percentage taken off the subtotal of a customer's first order.",
            "format": "double"
          },
          "deliveryFrom": {
            "type": "string",
            "description": "First day an order placed now should arrive on.",
            "format": "date"
          },
          "deliveryTo": {
            "type": "string",
            "description": "Last day an order placed now should arrive on.",
            "format": "date"
          }
        },
        "additionalProperties": false,
        "description": "The rules the order service prices an order by, so the cart page can preview the amounts\nthe same way instead of keeping its own copy of the numbers. The order itself is always\npriced here, at checkout."
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "StoreProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "traceId": {
            "type": "string",
            "description": "W3C trace id of the request, to find it in the logs.",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Messages per field of a 422 problem.",
            "nullable": true
          }
        },
        "additionalProperties": { },
        "description": "The error response as this store fills it, for the document only: the RFC 9457 members\nplus the trace id every problem carries and the field messages of a validation problem."
      },
...
      "WebhookReceipt": {
        "required": [
          "duplicate",
          "eventId"
        ],
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "duplicate": {
            "type": "boolean"
          }
        },
        "additionalProperties": false,
        "description": "The acknowledgement of a webhook: its id, and whether it had been received before."
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Access token issued by the identity service",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}